card12/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

pydantic>=2.8.0
pydantic-settings>=2.3.0

httpx==0.25.2
requests==2.31.0

sse-starlette==1.8.2

sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

loguru==0.7.2

python-dateutil==2.8.2

pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2


